Refactor node and evaluation classes
