--- operators/compute.py
@@ -1,6 +1,0 @@
-from typing import List
-
-
-def sqr(values: List[int]):
-    result = [val ** 2 for val in values]
-    return result
