+++ test/test_node.py
@@ -1,5 +1,8 @@
-from operators.node import Node, ModelNode, NodeFactory
-from operators.evaluation import LinRegression, LogRegression, XGBoost
+from core.evaluation import LinRegression, LogRegression, XGBoost
+from core.node import (
+    NodeFactory,
+    OperationNode
+)
 
 
 def test_node_log_reg():
@@ -10,6 +13,7 @@ def test_node_log_reg():
 def test_node_lin_log():
     test_node = NodeFactory().lin_reg()
     assert test_node.eval_strategy.__class__ == LinRegression
+    assert test_node.__class__ == OperationNode
 
 
 def test_node_xgboost():
